fix(typescript): move unrelated methods to appropriate services
